refactor(view-nfts): separate client logic
